requirements.txt
langchain
unstructured # Document loading
chromadb # Vector storage
openai # For embeddings
tiktoken  # For embeddings
pinecone-client
langchain
langchain-openai
langchain_community
langchain_pinecone
langchain[docarray]
docarray
python-dotenv
tiktoken 
pinecone-client 
scikit-learn
datasets==2.10.1
pypdf

weaviate-client
unstructured